requirements.txt
albumentations==1.3.0
matplotlib==3.6.3
numpy==1.21.2
opencv_python==4.7.0.72
pandas==1.5.2
Pillow==9.4.0
rich==13.3.3
torch==2.0.0
torchvision==0.15.1
